requirements-dev.txt
coverage
zappa
eveapimongo
grequests
evelink>=0.7.5
discord.py